+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Changed
-- Consecutive streaming outputs on `stdout` in a cell output context (vice widget context) will be "squashed" into a single output before passed to `.add_cell_content`
+- Add option in `OutputHandler` to squash streaming outputs on `stdout` in a cell output context (vice widget context) in order to create a single output before entering `.add_cell_content`
 
 ## [0.5.5] - 2023-04-26
 ### Fixed